Fixed several issues connected to mapping of markings.
Made all marking logic depend on the actual rendered pdf page instead of
a separate value which could come out of sync. Also made instantiation
of arrays for saving of markings either based on actual number of pages
in presentation, or dynamic assignment, so that it will scale to a
presentation of any size.